Move files from Configuration to Models
