--- Builder/sources/Domain/Configuration/TemplateLoopHtmlFile.cs
@@ -1,15 +1,0 @@
-﻿namespace Builder.Domain.Configuration
-{
-   public class TemplateLoopHtmlFile
-   {
-      public string BlogArchive { get; set; }
-      public string BlogCategories { get; set; }
-      public string BlogPostLast10 { get; set; }
-      public string BlogPostLast5 { get; set; }
-      public string BlogPostLast3 { get; set; }
-      public string BlogPosts { get; set; }
-      public string BlogTags { get; set; }
-      public string Menu { get; set; }
-      public string SocialMedia { get; set; }
-   }
-}
